cognitive-engine/requirements.txt
paho-mqtt
rtamt